requirements.txt
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rdp==0.8
six==1.16.0
